moved test into subfolder
