+++ src/components/RichTextEditor/__tests__/fixtures/selectors.ts
@@ -1,0 +1,22 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+export const TOOLBAR_PLUGIN_OL = '[data-plugin-id=ol]';
+export const RICH_TEXT_EDITOR = '[data-test-id=rich-text-editor]';
+export const TOOLBAR_FLOATING = '[data-test-id=toolbar-floating]';
+export const TOOLBAR_GROUP_1 = '[data-test-id=toolbar-group-1]';
+export const TOOLBAR_GROUP_2 = '[data-test-id=toolbar-group-2]';
+export const TEXTSTYLE_DROPDOWN_TRIGGER = '[data-test-id=textstyle-dropdown-trigger]';
+export const CHANGE_DESIGN_TOKENS_TRIGGER = '[data-test-id=change-design-tokens-button]';
+export const TEXTSTYLE_OPTION = '[data-test-id=textstyle-option]';
+export const CHECKBOX_INPUT = '[data-test-id=checkbox-input]';
+export const EDIT_LINK_BUTTON = '[data-test-id=edit-link-button]';
+export const EDIT_BUTTON_BUTTON = '[data-test-id=edit-button-button]';
+export const REMOVE_LINK_BUTTON = '[data-test-id=remove-link-button]';
+export const REMOVE_BUTTON_BUTTON = '[data-test-id=remove-button-button]';
+export const FLOATING_LINK_INSERT = '[data-test-id=floating-link-insert]';
+export const FLOATING_BUTTON_INSERT = '[data-test-id=floating-button-insert]';
+export const FLOATING_LINK_EDIT = '[data-test-id=floating-link-edit]';
+export const FLOATING_BUTTON_EDIT = '[data-test-id=floating-button-edit]';
+export const FLOATING_BUTTON_SECONDARY = '[data-test-id=floating-button-insert-secondary]';
+export const BUTTON = '[data-test-id=button]';
+export const CHECKBOX_INPUT_ID = '[data-test-id=checkbox-input]';
